+++ src/client/render/gl/passes/name-pass/IconProgram.ts
@@ -33,6 +33,7 @@ export class IconProgram {
 
   // Dynamic uniform locations
   private uCamera: WebGLUniformLocation;
+  private uScreenFacing: WebGLUniformLocation;
   private uTime: WebGLUniformLocation;
   private uLerpSpeed: WebGLUniformLocation;
   private uCullThreshold: WebGLUniformLocation;
@@ -93,6 +94,7 @@ export class IconProgram {
 
     // Dynamic uniform locations
     this.uCamera = gl.getUniformLocation(this.program, "uCamera")!;
+    this.uScreenFacing = gl.getUniformLocation(this.program, "uScreenFacing")!;
     this.uTime = gl.getUniformLocation(this.program, "uTime")!;
     this.uLerpSpeed = gl.getUniformLocation(this.program, "uLerpSpeed")!;
     this.uCullThreshold = gl.getUniformLocation(
@@ -149,6 +151,7 @@ export class IconProgram {
     settings: RenderSettings,
     vao: WebGLVertexArrayObject,
     fadeOwnerID: number,
+    screenFacing = false,
   ): void {
     if (!this.emojiReady) return;
 
@@ -157,6 +160,7 @@ export class IconProgram {
     gl.useProgram(this.program);
 
     gl.uniformMatrix3fv(this.uCamera, false, cameraMatrix);
+    gl.uniform1i(this.uScreenFacing, screenFacing ? 1 : 0);
     gl.uniform1f(this.uTime, performance.now() / 1000);
     gl.uniform1f(this.uLerpSpeed, ns.lerpSpeed);
     gl.uniform1f(this.uCullThreshold, ns.cullThreshold);