+++ src/client/render/gl/passes/name-pass/StatusIconProgram.ts
@@ -34,6 +34,7 @@ export class StatusIconProgram {
 
   // Dynamic uniform locations
   private uCamera: WebGLUniformLocation;
+  private uScreenFacing: WebGLUniformLocation;
   private uTime: WebGLUniformLocation;
   private uLerpSpeed: WebGLUniformLocation;
   private uCullThreshold: WebGLUniformLocation;
@@ -103,6 +104,7 @@ export class StatusIconProgram {
 
     // Dynamic uniform locations
     this.uCamera = gl.getUniformLocation(this.program, "uCamera")!;
+    this.uScreenFacing = gl.getUniformLocation(this.program, "uScreenFacing")!;
     this.uTime = gl.getUniformLocation(this.program, "uTime")!;
     this.uLerpSpeed = gl.getUniformLocation(this.program, "uLerpSpeed")!;
     this.uCullThreshold = gl.getUniformLocation(
@@ -159,6 +161,7 @@ export class StatusIconProgram {
     settings: RenderSettings,
     vao: WebGLVertexArrayObject,
     fadeOwnerID: number,
+    screenFacing = false,
   ): void {
     if (!this.atlasReady) return;
 
@@ -167,6 +170,7 @@ export class StatusIconProgram {
     gl.useProgram(this.program);
 
     gl.uniformMatrix3fv(this.uCamera, false, cameraMatrix);
+    gl.uniform1i(this.uScreenFacing, screenFacing ? 1 : 0);
     gl.uniform1f(this.uTime, performance.now() / 1000);
     gl.uniform1f(this.uLerpSpeed, ns.lerpSpeed);
     gl.uniform1f(this.uCullThreshold, ns.cullThreshold);